added stackAllocator and basic tests

+++ SirEngineThe3rdLib/src/SirEngine/globals.h
@@ -9,7 +9,7 @@ extern unsigned int SCREEN_WIDTH;
 extern unsigned int SCREEN_HEIGHT;
 extern Camera3DPivot *MAIN_CAMERA;
 extern GameClock GAME_CLOCK;
-extern uint32_t LAST_FRAME_TIME_NS;
+extern uint64_t LAST_FRAME_TIME_NS;
 extern uint32_t TOTAL_NUMBER_OF_FRAMES;
 #define FRAME_BUFFERS_COUNT 2
